Separate out API header
