+++ tests/ttnn/unit_tests/gtests/ccl/test_fabric_edm_common.hpp
@@ -11,7 +11,7 @@
 #include <tt-metalium/host_api.hpp>
 #include <tt-metalium/kernel.hpp>
 #include "tt-metalium/kernel_types.hpp"
-#include <tt-metalium/fabric_host_utils.hpp>
+#include <tt-metalium/fabric.hpp>
 #include "tt_metal/test_utils/df/df.hpp"
 #include "tt_metal/test_utils/env_vars.hpp"
 #include <tt-metalium/fabric_edm_packet_header.hpp>
