+++ tt_metal/fabric/CMakeLists.txt
@@ -8,6 +8,7 @@ target_sources(
         routing_table_generator.cpp
         mesh_graph.cpp
         erisc_datamover_builder.cpp
+        fabric.cpp
         fabric_host_utils.cpp
 )
